+++ Assets/Prefabs/Enemy.prefab
@@ -453,7 +453,7 @@ MonoBehaviour:
   m_PreserveAspect: 0
   m_FillCenter: 1
   m_FillMethod: 0
-  m_FillAmount: 0
+  m_FillAmount: 0.32
   m_FillClockwise: 1
   m_FillOrigin: 0
   m_UseSpriteMesh: 0
