+++ Assets/Scripts/AIBoats.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e6a444fa19625614eb4de7caaf8dba18
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
